--- components/item/Item.js
@@ -1,31 +1,0 @@
-import React, { Component } from "react";
-import { styles } from "./styles";
-import { Colors } from "../../colors/Colors";
-import { TouchableOpacity, Text, View } from "react-native";
-import { CheckBox } from "../checkbox/Checkbox";
-
-export class Item extends Component {
-  render() {
-    return (
-      <TouchableOpacity
-        style={{ backgroundColor: Colors.white }}
-        onPress={this.props._onPress}
-      >
-        <View style={styles.container}>
-          <View style={styles.textContainer}>
-            <Text style={styles.primaryText}>
-              {this.props["first-description"]}
-            </Text>
-            <Text style={styles.secondaryText}>
-              {this.props["second-description"]}
-            </Text>
-          </View>
-          <CheckBox
-            checked={this.props["status"]}
-            click={this.props.checkboxClick}
-          />
-        </View>
-      </TouchableOpacity>
-    );
-  }
-}
