+++ index.js
@@ -1,5 +1,5 @@
 import { AppRegistry } from "react-native";
-import { AppNavigator } from "./containers/home/Home";
+import { AppNavigator } from "./navigation/mainNavigation";
 import { name as appName } from "./app.json";
 
 AppRegistry.registerComponent(appName, () => AppNavigator);
